requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-docx==1.1.2
sentence-transformers==3.3.1
supabase==2.11.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
huggingface_hub>=0.25.0
redis==5.2.1
apscheduler>=3.10.0
